examples/fridge-object-detection/model_factory/common/smoke_test/modules/TestModule/requirements.txt
azure-iot-device~=2.7.0
json-schema-matcher == 0.1.7.1